% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckFormula.R
\name{CheckFormula}
\alias{CheckFormula}
\title{Check formula with provided data}
\usage{
CheckFormula(formula, y = NULL, groups = NULL, xName = "auto", yName = "auto")
}
\arguments{
\item{formula}{Formula to be used in model. If provided by user, formula
should include all values in yName and xName. Else, formula will be built
using provided data (see details for more information).}

\item{y}{A matrix containing values to be used on left-hand side of model
(e.g. dependent variable). At least one of y or groups must be provided.}

\item{groups}{A data.frame containing sample group data. At least one of y or
groups must be provided. Column one of groups data.frame must contain
sample names that correspond to column names of x and y.}

\item{xName}{Names of matrices provided in x. If x is a list with named
elements, defaults to list names. Else, defaults to x1, x2, ..., xn where n
is the number of matrices provided in x.}

\item{yName}{Name of matrix or name of column in groups to be used on
left-hand of model. (see formula details for default values)}
}
\value{
formula
}
\description{
\code{CheckFormula} is an internal function that checks formula if formula is
defined in ModelMultiData.
}
\details{
Checks that user defined formula include all values in yName and xName (e.g.
names of provided matrices).
}
