% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckFormulaLevels.R
\name{CheckFormulaLevels}
\alias{CheckFormulaLevels}
\title{Check formula for singular values}
\usage{
CheckFormulaLevels(formula, dt)
}
\arguments{
\item{formula}{Formula to be used in model.}

\item{dt}{data.table with model data.}
}
\value{
formula with singular values dropped.
}
\description{
\code{CheckFormulaLevels} is an internal function that tests that all
variables in formula have more than one value.
}
\details{
This function checks all variables in formula for singular values, which
usually cause errors when modeling, and drops singular variables.
}
