% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckGroups.R
\name{CheckGroups}
\alias{CheckGroups}
\title{Check Class of data provided to x parameter in ModelMultiData}
\usage{
CheckGroups(groups)
}
\arguments{
\item{groups}{data.frame or data.table with groups data}
}
\value{
data.table with groups data
}
\description{
\code{CheckGroups} is an internal function that checks the class of groups
parameter in ModelMultiData function.
}
