% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckReturnVars.R
\name{CheckReturnVars}
\alias{CheckReturnVars}
\title{Check that all variables in returnVars are included in formula}
\usage{
CheckReturnVars(returnVars, formula)
}
\arguments{
\item{returnVars}{Character vector of coefficient variable names to return.
By default, returns all coefficients for variables in xName. To return all
coefficients (including intercept), use "*".}

\item{formula}{Formula to be used in model. If provided by user, formula
should include all values in yName and xName. Else, formula will be built
using provided data (see details for more information).}
}
\description{
\code{returnVars} is an internal function that checks that all variables in
returnVars are included in formula in ModelMultiData.
}
