% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckXClass.R
\name{CheckXClass}
\alias{CheckXClass}
\title{Check Class of data provided to x parameter in ModelMultiData}
\usage{
CheckXClass(x)
}
\arguments{
\item{x}{matrix or list of matrices}
}
\value{
list of matrices
}
\description{
\code{CheckXClass} is an internal function that checks the class of x
parameter in ModelMultiData function.
}
