% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckYClass.R
\name{CheckYClass}
\alias{CheckYClass}
\title{Check Class of data provided to y parameter in ModelMultiData}
\usage{
CheckYClass(y)
}
\arguments{
\item{y}{A matrix containing values to be used on left-hand side of model
(e.g. dependent variable).}
}
\value{
matrix
}
\description{
\code{CheckYClass} is an internal function that checks the class of y
parameter in ModelMultiData function.
}
