% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce.R
\name{coalesce}
\alias{coalesce}
\title{Coalesce vectors with NAs}
\usage{
coalesce(...)
}
\arguments{
\item{...}{data.frame or list of vectors of equal length}
}
\value{
A vector of coalesced values
}
\description{
\code{coalesce} returns value which is not NA across columns of a data.frame
or a list of vectors
}
\examples{
dat <- list(c(1, NA, NA), c(NA, 2, NA), c(NA, NA, 3))
coalesce(dat)
dat <- data.frame(c(1, NA, NA), c(NA, 2, NA), c(NA, NA, 3))
coalesce(dat)
}
