% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSharedSamples.R
\name{GetSharedSamples}
\alias{GetSharedSamples}
\title{Find Shared Sample Names}
\usage{
GetSharedSamples(x, y = NULL, groups = NULL)
}
\arguments{
\item{x}{matrix or list of matrices}

\item{y}{matrix}

\item{groups}{data.frame or data.table with groups data}
}
\value{
vector of shared sample names
}
\description{
\code{GetSharedSamples} is an internal function that checks identifies names
that are shared in x, y, and the first column of groups as relevant within
ModelMultiData function.
}
