% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeIndividualFiles.R
\name{MergeIndividualFiles}
\alias{MergeIndividualFiles}
\title{Merge Data from Multiple Files}
\usage{
MergeIndividualFiles(
  fileDirectory,
  filePattern = NULL,
  filePrefix = NULL,
  indVars = "auto",
  IDcol = 1,
  ...
)
}
\arguments{
\item{fileDirectory}{a character vector of full path names to directory
containing individual files.}

\item{filePattern}{an optional regular expression. Only file names which
match the regular expression will be returned. Strings matching regular
expression will be removed from column names in returned dataset.}

\item{filePrefix}{an optional regular expression. Strings matching regular
expression will be removed from column names in returned dataset.}

\item{indVars}{character vector of column names in individual files that
correspond with individual measures (e.g. not annotation). Default value is
all columns except those in IDcol.}

\item{IDcol}{character or numeric vector of columns in individual files on
which to merge}

\item{...}{additional parameters to be passed to data.table::fread()}
}
\value{
A list of merged data. The first list element contains annotations
  (e.g. columns that are not included in indVars). All other list elements
  are matrices with rownames matching values in IDcol and columns matching
  file names.
}
\description{
\code{MergeIndividualFiles} merges data from individual sample data and
returns list of matrices.
}
\details{
This function reads in files containing data from individual samples and
merges individual data into one dataset by variable or measurement.
}
\examples{
\dontrun{
out <- MergeIndividualFiles(fileDirectory = "./sampleData/", filePattern =
".txt.gz", indVars = c("AllSubs", "Coverage-q25", "MeanQ",
"BaseCount[A,C,G,T]", "Frequency", "Pvalue"), IDcol = c("Region", "Position",
"Reference", "Strand"))
}
}
