% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAEdCombinations.R
\docType{data}
\name{RNAEdCombinations}
\alias{RNAEdCombinations}
\title{RNA Editing QTL Comparisons Matrix}
\format{1 matrix with 22 observations and 2 variables}
\usage{
RNAEdCombinations
}
\description{
A matrix with all comparisons between x1 and y to be tested in edQTL analysis
example.
}
\keyword{datasets}
