% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAEdData.R
\docType{data}
\name{RNAEdData}
\alias{RNAEdData}
\title{RNA Editing Data Object}
\format{1 list with 1 data.frame and 6 matrices.
\describe{
  \item{RNAEdData[["Annotation"]]}{A data.frame with 1000 observations and
  17 variables}
  \item{RNAEdData[["AllSubs"]]}{A matrix with 1000 observations and 8
  variables}
  \item{RNAEdData[["Coverage-q25"]]}{A matrix with 1000 observations and 8
  variables}
  \item{RNAEdData[["MeanQ"]]}{A matrix with 1000 observations and 8
  variables}
  \item{RNAEdData[["BaseCount[A,C,G,T]"]]}{A matrix with 1000 observations
  and 8 variables}
  \item{RNAEdData[["Frequency"]]}{A matrix with 1000 observations and 8
  variables}
  \item{RNAEdData[["Pvalue"]]}{A matrix with 1000 observations and 8
  variables}
}}
\usage{
RNAEdData
}
\description{
A list with 1 Annotation data.frame and 6 matrices containing RNA editing
data.
}
\keyword{datasets}
