% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAEdGenotypes.R
\docType{data}
\name{RNAEdGenotypes}
\alias{RNAEdGenotypes}
\title{Genotypes for RNA Editing QTL Analysis}
\format{1 matrix with 10 observations and 8 variables}
\usage{
RNAEdGenotypes
}
\description{
A matrix with genotypes at ten SNPs.
}
\keyword{datasets}
