% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAEdSampleInfo.R
\docType{data}
\name{RNAEdSampleInfo}
\alias{RNAEdSampleInfo}
\title{Sample Information for RNA Editing Data Sets}
\format{A data.frame with 8 observations and 4 variables.}
\usage{
RNAEdSampleInfo
}
\description{
A data.frame with sample names, age, status, and batch for each sample.
}
\keyword{datasets}
