% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunModel.R
\name{RunModel}
\alias{RunModel}
\title{Run regression model}
\usage{
RunModel(
  formula,
  modelData,
  comparison = NULL,
  returnVars = "*",
  FUN = lm,
  ...
)
}
\arguments{
\item{formula}{Formula to be used in model.}

\item{modelData}{data.table with all data to be used in model. Constructed
with BuildModelData function.}

\item{comparison}{Vector containing rownames or numbers of x and y included
in test.}

\item{returnVars}{Character vector of coefficient variable names to return.
By default, returns all coefficients for variables in xName. To return all
coefficients (including intercept), use "*".}

\item{FUN}{Function to be used for regression analysis. Defaults to lm().}

\item{...}{Additional parameters to be passed to FUN.}
}
\value{
A data.table with model coefficients for all variables in returnVars.
  Includes comparison, variable name, Estimate, Std. Error, t- or z-value,
  and pValue.
}
\description{
\code{RunModel} is an internal function that applies a regression model to
constructed model data and regurns model coefficients.
}
